backend/notification-service/requirements.txt
fastapi
strawberry-graphql
graphql-core
pymongo
boto3
uvicorn
python-dotenv
pydantic